# docs/openapi3_psy_help.yaml
openapi: 3.0.2
info:
  title: 'Psychology help'
  description: 'Документация API к сервису PSYHELP. [Актуальная версия на GitHub](https://github.com/py-psychological-help/psychological-help-backend/blob/develop/docs/openapi3_psy_help.yaml)'
  version: '0.1.5'
servers:
  - url: .../api/v1
paths:
...
  /chats/{chat_secret_key}/start/:
    parameters:
      - name: chat_secret_key
        required: true
        in: path
        description: A chat secret key.
        schema:
          type: string
    post:
      operationId: activate_chat
      description: 'Start chat by psychologist. При обращении к эндпоинту chat_secret_key отправляется клиенту на почту.'
      parameters: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
          description: ''
        '403':
            description: Forbidden.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Forbidden'
        '404':
            description: Not found.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Not_Found'
      tags:
      - api
...
components:
  schemas:
...
    Chat:
      type: object
      properties:
        id:
          type: integer
          format: integer
          readOnly: true
          example: 1122
        chat_secret_key:
          type: string
          format: string
          readOnly: true
        active:
          type: boolean
          format: boolian
          readOnly: true
        new:
          type: boolean
          format: boolian
          readOnly: true
        client:
          type: object
          properties:
            complaint:
              type: string
              format: string
            text:
              type: string
              format: string
            first_name:
              type: string
              readOnly: true
        psychologist:
          type: object
          properties:
            complaint:
              type: string
              format: string
            text:
              type: string
              format: string
            first_name:
              type: string
              readOnly: true
        messages:
          type: array
          readOnly: true
          items:
            type: object
            properties:
              id:
                type: integer
                readOnly: true
              date_time:
                type: string
              text:
                type: string
              is_author_me:
                type: boolean
                example: true
              author:
                type: integer
                example: 1
      required:
        - complaint
        - email
...
    Forbidden:
      type: object
      properties:
        detail:
            type: string
            example: "У вас недостаточно прав для выполнения данного действия."
    Not_Found:
      type: object
      properties:
        detail:
          type: string
          example: "Страница не найдена."